[timers] Make timers completely async/event driven (#1542)
Timers currently used the main event loop to check if
the timer had expired. This change uses Zephyr's timer
callbacks to signal the JS function. The linux port
uses signals to do the same thing.

+++ src/zjs_callbacks.c
@@ -302,7 +302,7 @@ zjs_callback_id add_callback_priv(jerry_value_t js_func,
     cb_map[new_cb->id] = new_cb;
 
     DBG_PRINT("adding new callback id %d, js_func=%p, once=%u\n", new_cb->id,
-              (void *)new_cb->js_func, once);
+            (void *)(uintptr_t)new_cb->js_func, once);
 
 #ifdef INSTRUMENT_CALLBACKS
     set_info_string(cb_map[new_cb->id]->creator, file, func);
@@ -485,7 +485,8 @@ void print_callbacks(void)
                 ZJS_PRINT("[%u] JS Callback:\n\tType: ", i);
                 if (jerry_value_is_function(cb_map[i]->js_func)) {
                     ZJS_PRINT("Single Function\n");
-                    ZJS_PRINT("\tjs_func: %p\n", (void *)cb_map[i]->js_func);
+                    ZJS_PRINT("\tjs_func: %p\n",
+                            (void *)(uintptr_t)cb_map[i]->js_func);
                     ZJS_PRINT("\tonce: %u\n", GET_ONCE(cb_map[i]->flags));
                 } else {
                     ZJS_PRINT("js_func is not a function\n");
